Revamp `exam 2`
