+++ src/main/kotlin/com/dicoding/exam/exam2/Exam2.kt
@@ -1,0 +1,11 @@
+package com.dicoding.exam.exam2
+
+// TODO 1
+fun calculate(valueA: Int, valueB: Int, valueC: Int?): Int {
+    return 0
+}
+
+// TODO 2
+fun result(result: Int): String {
+    return ""
+}
