+++ src/main/kotlin/com/dicoding/exam/exam2/Exam2Main.kt
@@ -1,0 +1,17 @@
+package com.dicoding.exam.exam2
+
+private fun main() {
+    val valueA = 101
+    val valueB = 52
+    val valueC = 99
+
+    val resultA = calculate(valueA, valueB, valueC)
+    val resultB = calculate(valueA, valueB, null)
+
+    println(
+        """
+        ${result(resultA)}
+        ${result(resultB)}
+        """.trimIndent()
+    )
+}
